backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
alembic==1.13.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.2
loguru==0.7.2
slowapi==0.1.9
starlette==0.38.5
jinja2==3.1.4
reportlab==4.2.2
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
torch==2.4.1
shap==0.46.0
matplotlib==3.9.2
joblib==1.4.2
langchain==0.2.16
passlib==1.7.4